cause() works for errors from different contexts
err instanceof Error does not work with errors from different contexts
because the prototype chain is different (e.g. Error in context 1 is
different from Error in context 2). The node core gets around this by
using isError, which checks for errors in the following way:

```
function isError(e) {
  return (objectToString(e) === '[object Error]' || e instanceof Error);
}
```

Which is backwards compatible with the current approach of only checking
`instanceof Error`.

 - Add tst.context.js to test that 'cause()' works for errors from
   different contexts.
 - Add 'core-util-is'.
 - Replace all instances of 'instanceof Error' with 'mod_isError(err)'.